nav bar and router
