+++ .~lock.f1-plan slides.odp#
@@ -1,0 +1,1 @@
+,DESKTOP-LGHDMNA/Pc,DESKTOP-LGHDMNA,22.03.2021 16:40,file:///C:/Users/Pc/AppData/Roaming/LibreOffice/4;
