+++ src/components/Race.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Race = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Race
