+++ src/components/Team.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Team = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Team
