+++ src/index.js
@@ -3,13 +3,16 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { BrowserRouter as Router } from 'react-router-dom'
 
 
 
 
 ReactDOM.render(
   <React.StrictMode>
+    <Router>
       <App />
+    </Router>
   </React.StrictMode>,
   document.getElementById('root')
 );
